+++ Emulator/Cartridges/CustomCartridges/Expert.h
@@ -93,9 +93,9 @@ private:
     // Operating buttons
     //
     
-    long numButtons() const override { return 2; }
-    const char *getButtonTitle(unsigned nr) const override;
-    void pressButton(unsigned nr) override;
+    isize numButtons() const override { return 2; }
+    const char *getButtonTitle(isize nr) const override;
+    void pressButton(isize nr) override;
     
     
     //
