+++ Emulator/Cartridges/CustomCartridges/FinalIII.h
@@ -115,8 +115,8 @@ private:
     // Operating buttons
     //
     
-    long numButtons() const override { return 2; }
-    const char *getButtonTitle(unsigned nr) const override;
-    void pressButton(unsigned nr) override;
-    void releaseButton(unsigned nr) override;
+    isize numButtons() const override { return 2; }
+    const char *getButtonTitle(isize nr) const override;
+    void pressButton(isize nr) override;
+    void releaseButton(isize nr) override;
 };
