+++ Emulator/Cartridges/CustomCartridges/GameKiller.cpp
@@ -46,13 +46,13 @@ GameKiller::pokeIO2(u16 addr, u8 value)
 }
 
 const char *
-GameKiller::getButtonTitle(unsigned nr) const
+GameKiller::getButtonTitle(isize nr) const
 {
     return nr == 1 ? "Freeze" : nullptr;
 }
 
 void
-GameKiller::pressButton(unsigned nr)
+GameKiller::pressButton(isize nr)
 {
     if (nr == 1) {
         c64.signalExpPortNmi();
@@ -60,7 +60,7 @@ GameKiller::pressButton(unsigned nr)
 }
 
 void
-GameKiller::releaseButton(unsigned nr)
+GameKiller::releaseButton(isize nr)
 {
     if (nr == 1) {
         suspend();
