+++ Emulator/Cartridges/CustomCartridges/WarpSpeed.cpp
@@ -53,13 +53,13 @@ WarpSpeed::pokeIO2(u16 addr, u8 value)
 }
 
 const char *
-WarpSpeed::getButtonTitle(unsigned nr) const
+WarpSpeed::getButtonTitle(isize nr) const
 {
     return nr == 1 ? "Reset" : nullptr;
 }
 
 void
-WarpSpeed::pressButton(unsigned nr)
+WarpSpeed::pressButton(isize nr)
 {
     assert(nr <= numButtons());
